app/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
passlib
python-jose
python-dotenv
pydantic_settings
pydantic
PyJWT
psycopg2-binary